Propagate PRINTF_ALIAS_STANDARD_FUNCTION_NAMES
PRINTF_ALIAS_STANDARD_FUNCTION_NAMES is used only from the printf.h
header file, before the printf_config.h is included. It means that the
CMake option PRINTF_ALIAS_STANDARD_FUNCTION_NAMES will never be used.
This commit changes that, and propagates it as a compile definition,
when it is defined.